IHEC-CODELAB/requirements.txt
requests
beautifulsoup4
lxml
pandas
sqlalchemy
fastapi
uvicorn
python-dotenv
